Add distrobox awareness to vim-tmux-navigator
TL;DR -- the "ps" command vim-tmux-navigator uses to query processes
owned by the current tmux pane is useless if we're in the distrobox
shell, because the ps command runs outside of the container.

The integration here is tight but in essence we create a std path at
$XDG_RUNTIME_DIR/dbx-tmux, in which each pane hosting a distrobox shell
records its Container ID and tty within the container. Recording those
parameters has to be done on shell entry, so it lives in bashrc.

The in-container tty + the container ID can be used to find processes
within the proper cgroup and pseudoterminal combination to apply the
same check as vim-tmux-navigator does on the host.